// frontend/src/sections/purchase-orders/list/createPurchaseOrder.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InsertPurchaseOrderVariables = {
  storeId: string;
  input: {
    id: string;
    supplierId: string;
  };
};

export type InsertPurchaseOrderResult = {
  insertPurchaseOrder: ({
  __typename: "IdResponse";
} & {
  id: string;
});
};

export const InsertPurchaseOrder = {
  query: "mutation insertPurchaseOrder($storeId: String!, $input: InsertPurchaseOrderInput!) {\n  insertPurchaseOrder(storeId: $storeId, input: $input) {\n    __typename\n    ... on IdResponse {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertPurchaseOrderResult, InsertPurchaseOrderVariables>;